Oracle service provide call added

+++ contracts/src/main/kotlin/com/template/states/CashState.kt
@@ -6,7 +6,7 @@ import net.corda.core.contracts.ContractState
 import net.corda.core.identity.AbstractParty
 
 @BelongsToContract(CashContract::class)
-data class CashState(val exRate: Long, val transferor: AccountDetails, val transferee: AccountDetails,
+data class CashState(val exRate: Double, val transferor: AccountDetails, val transferee: AccountDetails,
                      override val participants: List<AbstractParty> = listOf(transferor.party, transferee.party)
 ) : ContractState
